fix: 리뷰관리 페이지 공개비공개 모달에서 책제목보이기 (#502)
* refactor: ReviewList 대신 ReviewListItem 으로 변경

* refactor: 불필요한 속성 할당 및 파싱 코드 제거

* refactor: 함수이름 변경 및 제네릭 제거

1. onClick -> requestConfirmToUpdate
함수의 동작을 유추할 수 있는 네이밍으로 변경

2. 불필요한 제네릭 타입 제거
가독성을 위해 타입 생략

* feat: 리뷰 공개비공개 수정시 책 제목도 보이도록
